Added Authentication and Authorization

+++ src/main/java/com/project/airBnb/airbnbApp/services/HotelServiceImpl.java
@@ -5,13 +5,16 @@ import com.project.airBnb.airbnbApp.dto.HotelInfoDto;
 import com.project.airBnb.airbnbApp.dto.RoomDto;
 import com.project.airBnb.airbnbApp.entity.Hotel;
 import com.project.airBnb.airbnbApp.entity.Room;
+import com.project.airBnb.airbnbApp.entity.User;
 import com.project.airBnb.airbnbApp.exceptions.ResourceNotFoundException;
+import com.project.airBnb.airbnbApp.exceptions.UnauthorisedException;
 import com.project.airBnb.airbnbApp.repositories.HotelRepository;
 import com.project.airBnb.airbnbApp.repositories.RoomRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -30,6 +33,10 @@ public class HotelServiceImpl implements HotelService{
         log.info("Creating a new hotel with name: {}", hotelDto.getName());
         Hotel hotel= modelMapper.map(hotelDto,Hotel.class);
         hotel.setActive(false);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        hotel.setOwner(user);
+
         hotel=hotelRepository.save(hotel);
         log.info("Created a new hotel with ID: {}", hotelDto.getId());
         return modelMapper.map(hotel,HotelDto.class);
@@ -41,6 +48,11 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel=hotelRepository
                 .findById(id)
                 .orElseThrow(()->new ResourceNotFoundException("Hotel not found with id "+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnauthorisedException("This user does not own this hotel with id: "+id);
+        }
         return modelMapper.map(hotel,HotelDto.class);
     }
 
@@ -49,6 +61,10 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id "+id));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnauthorisedException("This user does not own this hotel with id: "+id);
+        }
         modelMapper.map(hotelDto,hotel);
         hotel.setId(id);
         hotel = hotelRepository.save(hotel);
@@ -61,6 +77,12 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id "+id));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnauthorisedException("This user does not own this hotel with id: "+id);
+        }
+
         for(Room room: hotel.getRooms()) {
             inventoryService.deleteAllInventories(room);
             roomRepository.deleteById(room.getId());
@@ -74,6 +96,10 @@ public class HotelServiceImpl implements HotelService{
                 .findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id "+id));
 
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnauthorisedException("This user does not own this hotel with id: "+id);
+        }
         hotel.setActive(true);
 
         // assuming doing only once
@@ -88,6 +114,8 @@ public class HotelServiceImpl implements HotelService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id "+hotelId));
+
+
         List<RoomDto> rooms =hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element,RoomDto.class))
